requirements.txt
Flask==2.3.3
gspread==5.12.0
oauth2client==4.1.3
requests==2.31.0
python-dotenv==1.0.0
flask-session==0.5.0
flask-mail==0.9.1